+++ src/Method/Sticker/SetCustomEmojiStickerSetThumbnail.php
@@ -1,0 +1,48 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method\Sticker;
+
+use Vjik\TelegramBot\Api\ParseResult\ValueHelper;
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+
+/**
+ * @see https://core.telegram.org/bots/api#setcustomemojistickersetthumbnail
+ */
+final readonly class SetCustomEmojiStickerSetThumbnail implements TelegramRequestWithResultPreparingInterface
+{
+    public function __construct(
+        private string $name,
+        private ?string $customEmojiId = null,
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'setCustomEmojiStickerSetThumbnail';
+    }
+
+    public function getData(): array
+    {
+        return array_filter(
+            [
+                'name' => $this->name,
+                'custom_emoji_id' => $this->customEmojiId,
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
+    }
+
+    public function prepareResult(mixed $result): true
+    {
+        ValueHelper::assertTrueResult($result);
+        return $result;
+    }
+}