+++ CHANGELOG.md
@@ -44,6 +44,7 @@
   methods.
 - New #60: Add `setStickerSetTitle` method.
 - New #61: Add `setStickerSetThumbnail` method.
+- New #62: Add `setCustomEmojiStickerSetThumbnail` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.